Reusing js code for metabolite explorer views
